ch: create sendErrorMessageForUnsupportedMedia and separate codes
